refactor: move storage related methods to StorageManager class
